% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HealthServicesCosts.R
\name{calculateHealthCost}
\alias{calculateHealthCost}
\title{Function that will return default values of the program change values.}
\usage{
calculateHealthCost(
  UnitCosts = HealthServiceUnitCosts(),
  TbCases,
  LtbiTests,
  IGRA_frc = 0.5,
  discount = 0.3,
  tx_dist = c(1/3, 1/3, 1/3),
  LtbiTxInitsm,
  Years
)
}
\value{

}
\description{
Function that will return default values of the program change values.
}
