% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductivityCosts.R
\name{calculateProdCost}
\alias{calculateProdCost}
\title{Function that will return default values of the program change values.}
\usage{
calculateProdCost(
  TbCases,
  TbDeaths,
  LtbiTxInits,
  discount = 0,
  tx_dist = c(1/3, 1/3, 1/3),
  Ages = "SingleYearAge"
)
}
\arguments{
\item{TbCases}{a matrix of tuberculosis cases by year and age group}

\item{TbDeaths}{a matrix of tuberculosis deaths by year and age group}

\item{LtbiTxInits}{a matrix of ltbi treatment initations cases by year and age group}
}
\value{

}
\description{
Function that will return default values of the program change values.
}
