% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QALYCalculations.R
\name{calculateQALYs}
\alias{calculateQALYs}
\title{Function that will return default values of the program change values.}
\usage{
calculateQALYs(
  TbCases,
  TbDeaths,
  LtbiTxInits,
  discount = 0.03,
  Ages = "SingleYearAge"
)
}
\value{

}
\description{
Function that will return default values of the program change values.
}
