% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyCasesAverted.R
\name{dummyCases}
\alias{dummyCases}
\title{Function that will return a matrix of dummy values for cases averted}
\usage{
dummyCases(startyear = 2020, endyear = 2025, startage = 0, endage = 100)
}
\value{

}
\description{
Function that will return a matrix of dummy values for cases averted
}
