% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HealthServiceUnitCosts.R
\name{HealthServiceUnitCosts}
\alias{HealthServiceUnitCosts}
\title{Function that will return default values of the health services
costs associated with LTBI and TB in the United States. These values
are all represented in 2020 USD.}
\usage{
HealthServiceUnitCosts()
}
\value{
vector of the default values
}
\description{
Function that will return default values of the health services
costs associated with LTBI and TB in the United States. These values
are all represented in 2020 USD.
}
