% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TbVisuals.R
\name{TbVisuals}
\alias{TbVisuals}
\title{Function that will return different plots based on the arguments given}
\usage{
TbVisuals()
}
\arguments{
\item{TbCasesAverted}{a matrix of tuberculosis cases by year and age group}

\item{TbDeathsAverted}{a matrix of tuberculosis deaths by year and age group}

\item{LtbiTxInitsInc}{a matrix of ltbi treatment initations cases by year and age group}
}
\value{

}
\description{
Function that will return different plots based on the arguments given
}
